--- app/src/main/java/com/geekbrains/nasaexplorer/api/NasaApi.kt
@@ -1,19 +1,0 @@
-package com.geekbrains.nasaexplorer.api
-
-import com.geekbrains.nasaexplorer.api.apod.PictureOfTheDayResponse
-import com.geekbrains.nasaexplorer.api.asteroids_neows.AsteroidsNeoWSResponse
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-private const val NASA_APOD_URL = "planetary/apod"
-private const val NASA_ASTEROIDS_NEOWS_URL = "neo/rest/v1/feed"
-private const val API_KEY = "api_key"
-
-interface NasaApi {
-
-    @GET(NASA_APOD_URL)
-    suspend fun pictureOfTheDay(@Query(API_KEY) key: String): PictureOfTheDayResponse
-
-    @GET(NASA_ASTEROIDS_NEOWS_URL)
-    suspend fun asteroidsInfo(@Query(API_KEY) key: String): AsteroidsNeoWSResponse
-}
