--- app/src/main/java/com/geekbrains/nasaexplorer/domain/NasaRepository.kt
@@ -1,9 +1,0 @@
-package com.geekbrains.nasaexplorer.domain
-
-import com.geekbrains.nasaexplorer.api.apod.PictureOfTheDayInfo
-import com.geekbrains.nasaexplorer.api.asteroids_neows.AsteroidsNeoWSResponse
-
-interface NasaRepository {
-    suspend fun pictureOfTheDay(): PictureOfTheDayInfo
-    suspend fun asteroidsInfo(): AsteroidsNeoWSResponse
-}
