+++ app/src/main/java/com/stellarworker/nasaexplorer/domain/NasaRepository.kt
@@ -1,0 +1,9 @@
+package com.stellarworker.nasaexplorer.domain
+
+import com.stellarworker.nasaexplorer.api.apod.PictureOfTheDayInfo
+import com.stellarworker.nasaexplorer.api.asteroids_neows.AsteroidsNeoWSResponse
+
+interface NasaRepository {
+    suspend fun pictureOfTheDay(): PictureOfTheDayInfo
+    suspend fun asteroidsInfo(): AsteroidsNeoWSResponse
+}
