+++ app/src/main/java/com/stellarworker/nasaexplorer/domain/Note.kt
@@ -1,0 +1,8 @@
+package com.stellarworker.nasaexplorer.domain
+
+private const val EMPTY_STRING = ""
+
+data class Note(
+    var text: String = EMPTY_STRING,
+    var addedByUser: Boolean = false
+)
