+++ app/src/main/java/com/stellarworker/nasaexplorer/domain/Notes.kt
@@ -1,0 +1,6 @@
+package com.stellarworker.nasaexplorer.domain
+
+
+data class Notes(
+    var noteList: MutableList<Note> = mutableListOf()
+)
