+++ app/src/main/java/com/stellarworker/nasaexplorer/ui/MainFragmentDataset.kt
@@ -1,0 +1,9 @@
+package com.stellarworker.nasaexplorer.ui
+
+private const val EMPTY_STRING = ""
+
+data class MainFragmentDataset(
+    val image: String = EMPTY_STRING,
+    val title: String = EMPTY_STRING,
+    val explanation: String = EMPTY_STRING,
+)
