+++ app/src/main/java/com/stellarworker/nasaexplorer/utils/ViewUtils.kt
@@ -1,0 +1,40 @@
+package com.stellarworker.nasaexplorer.utils
+
+import android.view.View
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.google.android.material.snackbar.Snackbar
+
+private const val EMPTY_STRING = ""
+
+fun View.makeSnackbar(
+    text: String = EMPTY_STRING,
+    actionText: String = EMPTY_STRING,
+    action: (View) -> Unit = {},
+    length: Int = Snackbar.LENGTH_LONG,
+    anchor: View? = null
+) {
+    Snackbar
+        .make(this, text, length)
+        .setAction(actionText, action)
+        .setAnchorView(anchor)
+        .show()
+}
+
+fun View.show() {
+    visibility = View.VISIBLE
+}
+
+fun View.hide() {
+    visibility = View.GONE
+}
+
+fun View.hideKeyboard() =
+    ViewCompat.getWindowInsetsController(this)?.hide(WindowInsetsCompat.Type.ime())
+
+fun View.showKeyboard() =
+    ViewCompat.getWindowInsetsController(this)?.show(WindowInsetsCompat.Type.ime())
+
+inline fun View.visibleIf(condition: () -> Boolean) {
+    visibility = if (condition.invoke()) View.VISIBLE else View.GONE
+}
