+++ app/src/main/res/drawable/ic_twotone_note_add_24.xml
@@ -1,0 +1,8 @@
+<vector android:autoMirrored="true" android:height="24dp"
+    android:tint="#000000" android:viewportHeight="24"
+    android:viewportWidth="24" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillAlpha="0.3"
+        android:fillColor="@android:color/white"
+        android:pathData="M13,4L6,4v16h12L18,9h-5L13,4zM16,14v2h-3v3h-2v-3L8,16v-2h3v-3h2v3h3z" android:strokeAlpha="0.3"/>
+    <path android:fillColor="@android:color/white" android:pathData="M13,11h-2v3L8,14v2h3v3h2v-3h3v-2h-3zM14,2L6,2c-1.1,0 -2,0.9 -2,2v16c0,1.1 0.89,2 1.99,2L18,22c1.1,0 2,-0.9 2,-2L20,8l-6,-6zM18,20L6,20L6,4h7v5h5v11z"/>
+</vector>
