+++ app/src/main/res/layout/fragment_notes_recycler_item.xml
@@ -1,0 +1,82 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/fragment_notes_recycler_item_root"
+    style="@style/Background"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginHorizontal="@dimen/fragment_notes_recycler_item_card_margin_horizontal"
+    android:layout_marginTop="@dimen/fragment_notes_recycler_item_card_margin_top"
+    app:cardCornerRadius="@dimen/fragment_notes_recycler_item_card_corner_radius"
+    app:cardElevation="@dimen/fragment_notes_recycler_item_card_elevation">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_margin="@dimen/common_margin_horizontal">
+
+        <com.google.android.material.textfield.TextInputLayout
+            android:id="@+id/fragment_notes_recycler_item_note_text_container"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
+
+            <com.google.android.material.textfield.TextInputEditText
+                android:id="@+id/fragment_notes_recycler_item_note_text"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:inputType="textNoSuggestions|textMultiLine"
+                android:minHeight="20dp" />
+
+        </com.google.android.material.textfield.TextInputLayout>
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/fragment_notes_recycler_item_note_up"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/common_margin_horizontal"
+            android:layout_marginTop="@dimen/common_margin_horizontal"
+            android:src="@drawable/ic_baseline_arrow_circle_up_24"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/fragment_notes_recycler_item_note_down"
+            app:layout_constraintTop_toBottomOf="@id/fragment_notes_recycler_item_note_text_container" />
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/fragment_notes_recycler_item_note_down"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/common_margin_horizontal"
+            android:layout_marginTop="@dimen/common_margin_horizontal"
+            android:src="@drawable/ic_baseline_arrow_circle_down_24"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/fragment_notes_recycler_item_note_save"
+            app:layout_constraintTop_toBottomOf="@id/fragment_notes_recycler_item_note_text_container" />
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/fragment_notes_recycler_item_note_save"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/common_margin_horizontal"
+            android:layout_marginTop="@dimen/common_margin_horizontal"
+            android:src="@drawable/ic_baseline_save_24"
+            android:visibility="gone"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/fragment_notes_recycler_item_note_delete"
+            app:layout_constraintTop_toBottomOf="@id/fragment_notes_recycler_item_note_text_container" />
+
+        <androidx.appcompat.widget.AppCompatImageView
+            android:id="@+id/fragment_notes_recycler_item_note_delete"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginHorizontal="@dimen/common_margin_horizontal"
+            android:layout_marginTop="@dimen/common_margin_horizontal"
+            android:src="@drawable/ic_baseline_delete_24"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/fragment_notes_recycler_item_note_text_container" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.cardview.widget.CardView>
