+++ app/src/main/res/menu/bottom_navigation.xml
@@ -7,8 +7,13 @@
 
     <item
         android:id="@+id/navigation_asteroids"
-        android:icon="@drawable/ic_baseline_list_alt_24"
+        android:icon="@drawable/ic_baseline_circle_24"
         android:title="@string/bottom_navigation_asteroids" />
+
+    <item
+        android:id="@+id/navigation_notes"
+        android:icon="@drawable/ic_baseline_notes_24"
+        android:title="@string/bottom_navigation_notes" />
 
     <item
         android:id="@+id/navigation_settings"
