+++ app/src/main/res/menu/notes_fragment_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/create_note"
+        android:icon="@drawable/ic_baseline_note_add_24"
+        android:title="@string/notes_fragment_menu_create_note"
+        app:showAsAction="never" />
+</menu>
