+++ app/src/main/res/values/dimens.xml
@@ -29,4 +29,15 @@
     <dimen name="fragment_asteroidsneows_recycler_item_card_elevation">5dp</dimen>
     <!-- AsteroidsNeoWSFragment RecyclerItem end -->
 
+    <!-- NotesFragment begin -->
+
+    <!-- NotesFragment end -->
+
+    <!-- NotesFragment RecyclerItem begin -->
+    <dimen name="fragment_notes_recycler_item_card_margin_horizontal">5dp</dimen>
+    <dimen name="fragment_notes_recycler_item_card_margin_top">5dp</dimen>
+    <dimen name="fragment_notes_recycler_item_card_corner_radius">5dp</dimen>
+    <dimen name="fragment_notes_recycler_item_card_elevation">5dp</dimen>
+    <!-- NotesFragment RecyclerItem end -->
+
 </resources>
