+++ app/src/main/res/values/strings.xml
@@ -10,6 +10,7 @@
     <string name="bottom_navigation_home">Home</string>
     <string name="bottom_navigation_asteroids">Asteroids</string>
     <string name="bottom_navigation_settings">Settings</string>
+    <string name="bottom_navigation_notes">Notes</string>
     <!-- BottomNavigation end -->
 
     <!-- MainFragment begin -->
@@ -49,5 +50,14 @@
     <string name="fragment_asteroidsneows_recycler_item_approach_date">
         Close approach date: %1$s</string>
     <!-- AsteroidsNeoWSFragment RecyclerItem end -->
+
+    <!-- NotesFragment begin -->
+    <string name="notes_fragment_note_saved">Note Saved!</string>
+    <string name="notes_fragment_note_deleted">Note deleted!</string>
+    <!-- NotesFragment end -->
+
+    <!-- NotesFragment Menu begin -->
+    <string name="notes_fragment_menu_create_note">Create note</string>
+    <!-- NotesFragment Menu end -->
 
 </resources>
